# internal/model_new_submission.go
/*
 * MyHwProj
 *
 * REST api for MyHwProj
 *
 * API version: 1.0.0
 * Contact: myhwproj@yandex.ru
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package myhwproj

type NewSubmission struct {

	// Link to the solution
	Solution string `json:"solution"`
}

// AssertNewSubmissionRequired checks if the required fields are not zero-ed
func AssertNewSubmissionRequired(obj NewSubmission) error {
	elements := map[string]interface{}{
		"solution": obj.Solution,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertRecurseNewSubmissionRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of NewSubmission (e.g. [][]NewSubmission), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseNewSubmissionRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aNewSubmission, ok := obj.(NewSubmission)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertNewSubmissionRequired(aNewSubmission)
	})
}